requirements.txt
# ACME Store - Python Worker Service dependencies
# Pinned to specific versions for SCA testing
# NOTE: deliberately uses vulnerable versions for Sentinel SCA testing

# Web framework
Django==3.2.12

# HTTP client
requests==2.27.0

# Image processing
Pillow==9.0.0

# Data processing
numpy==1.23.1
pandas==1.4.2

# Database ORM
SQLAlchemy==1.4.40

# Task queue
celery==5.2.7
redis==4.3.4

# Utilities
python-dateutil==2.8.2
pytz==2022.1
pyyaml==5.4.1

# Dev/test
pytest==7.2.0
pytest-django==4.5.2
coverage==6.5.0
